HBX-1970: Move functionality of 'DatabaseReader#readSequences(String sql)' to new class 'org.hibernate.tool.internal.export.lint.SequenceCollector'
