+++ orm/src/main/java/org/hibernate/tool/internal/reveng/reader/DatabaseReader.java
@@ -151,42 +151,6 @@ public class DatabaseReader {
 
 	}
 
-	public Set<String> readSequences(String sql) {
-		Set<String> sequences = new HashSet<String>();
-		if (sql != null) {
-			Connection connection = null;
-			try {
-
-				connection = provider.getConnection();
-				Statement statement = null;
-				ResultSet rs = null;
-				try {
-					statement = connection.createStatement();
-					rs = statement.executeQuery(sql);
-					while (rs.next()) {
-						sequences.add(rs.getString("SEQUENCE_NAME").toLowerCase().trim());
-					}
-				} finally {
-					if (rs != null)
-						rs.close();
-					if (statement != null)
-						statement.close();
-				}
-
-			} catch (SQLException e) {
-				throw new RuntimeException("Problem while closing connection", e);
-			} finally {
-				if (connection != null)
-					try {
-						provider.closeConnection(connection);
-					} catch (SQLException e) {
-						throw new RuntimeException("Problem while closing connection", e);
-					}
-			}
-		}
-		return sequences;
-	}
-
 	private List<SchemaSelection> getSchemaSelections(String catalog, String schema) {
 		List<SchemaSelection> result = revengStrategy.getSchemaSelections();
 		if (result == null) {
